refactor(runtime): relocate memory isolation tracer
