+++ runtime/validation/__init__.py
@@ -1,0 +1,1 @@
+"""Runtime validation helpers with stable top-level compatibility modules."""
